Accident alert
Developed an accident alert system using ESP32 to track the
location of incidents. The system is designed to automatically send
GPS coordinates in case of an accident, allowing for rapid response.
The project involved programming the ESP32 for real-time data
processing and designing the circuit using Fritzing software. The
system was implemented and tested for functionality.